+++ NEWS.md
@@ -1,3 +1,8 @@
+# 2026-04-09
+
+- Add `predict_proba` to class `Model`
+- Add functions to compute Kernel Shap
+
 # 2026-04-03
 
 - Remove `type` from predict